requirements.txt
git+https://github.com/TeammyTurner/tablutpy.git@obs-rewards
gym>=0.17.3
stable-baselines3>=0.10.0